mqtt_client now works with awsiot.py library in testing.
